Refactor MakeTLSConfig and support mTLS on server (#739)
This change includes refactoring for MakeTLSConfig and adds
tests to ensure that we keep the current behavior. To make testing
easier, we pass os.ReadFile as an argument and use fstest.MapFS.ReadFile
in tests. In particular, this also means avoiding tls.LoadX509KeyPair.
Also, support for enabling mutual TLS authentication for Centrifugo server
was added.